+++ src/Deedle/AssemblyInfo.fs
@@ -4,10 +4,10 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("Deedle")>]
 [<assembly: AssemblyProductAttribute("Spreads")>]
 [<assembly: AssemblyDescriptionAttribute("Spreads")>]
-[<assembly: AssemblyVersionAttribute("0.0.57")>]
-[<assembly: AssemblyFileVersionAttribute("0.0.57")>]
+[<assembly: AssemblyVersionAttribute("0.0.60")>]
+[<assembly: AssemblyFileVersionAttribute("0.0.60")>]
 [<assembly: AssemblyCopyrightAttribute("(c) Victor Baybekov 2015")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "0.0.57"
+    let [<Literal>] Version = "0.0.60"